backend/pyproject.toml
[tool.poetry]
name = "pokeapp"
version = "0.1.0"
description = ""
authors = ["filpss <filipepinho2010@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = {extras = ["standard"], version = "^0.112.2"}
sqlalchemy = "^2.0.32"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.2"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.9"
pyjwt = "^2.9.0"

[tool.taskipy.tasks]
run = 'fastapi dev pokeapp/app.py'
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
pre_test = 'task lint'
test = 'pytest -s -x --cov=pokeapp -vv'
post_test = 'coverage html'
debug = 'uvicorn pokeapp.app:app --reload'
uvi = 'uvicorn pokeapp.app:app --host 0.0.0.0 --port 8000'

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.4.10"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
